# swagger.json
{"openapi":"3.0.1","info":{"title":"City Management API","description":"API documentation for the City Management application.","version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Authentication","description":"APIs for user authentication and account management"},{"name":"Service Management","description":"APIs for managing services within a city"},{"name":"Service Category Management","description":"APIs for managing service categories"},{"name":"City Management","description":"APIs for managing cities"}],"paths":{"/api/service-categories/{categoryId}":{"get":{"tags":["Service Category Management"],"summary":"Get Service Category by ID","description":"Retrieves a service category by its ID.","operationId":"getServiceCategoryById","parameters":[{"name":"categoryId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ServiceCategoryResponse"}}}}}},"put":{"tags":["Service Category Management"],"summary":"Update Service Category","description":"Updates an existing service category by its ID. Only admins can perform this action.","operationId":"updateServiceCategory","parameters":[{"name":"categoryId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServiceCategoryRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ServiceCategoryResponse"}}}}}},"delete":{"tags":["Service Category Management"],"summary":"Delete Service Category","description":"Deletes a service category by its ID. Only admins can perform this action.","operationId":"deleteServiceCategory","parameters":[{"name":"categoryId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/cities/{id}":{"get":{"tags":["City Management"],"summary":"Retrieve a city by ID","description":"Fetches the details of a specific city by its ID.","operationId":"getCityById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CityResponse"}}}}}},"put":{"tags":["City Management"],"summary":"Update an existing city","description":"Updates the details of an existing city by its ID. Only admins can perform this action.","operationId":"updateCity","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CityRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CityResponse"}}}}}},"delete":{"tags":["City Management"],"summary":"Delete a city","description":"Deletes a city by its ID. Only admins can perform this action.","operationId":"deleteCity","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/account":{"get":{"tags":["Authentication"],"summary":"Get current user","description":"Retrieves the details of the currently authenticated user.","operationId":"getCurrentUser","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GetCurrentUserResponse"}}}}}},"put":{"tags":["Authentication"],"summary":"Edit account","description":"Edits the current user's account details.","operationId":"editAccount","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EditAccountRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GetCurrentUserResponse"}}}}}}},"/api/account/role":{"put":{"tags":["Authentication"],"summary":"Change User Role","description":"Allows an admin to change the role of a user.","operationId":"changeUserRole","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangeUserRoleRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/service-categories":{"get":{"tags":["Service Category Management"],"summary":"Get All Service Categories","description":"Retrieves all service categories.","operationId":"getAllServiceCategories","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ServiceCategoryResponse"}}}}}}},"post":{"tags":["Service Category Management"],"summary":"Create Service Category","description":"Creates a new service category. Only admins can perform this action.","operationId":"createServiceCategory","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServiceCategoryRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ServiceCategoryResponse"}}}}}}},"/api/cities/{cityId}/services":{"get":{"tags":["Service Management"],"summary":"Get all services in a city","description":"Retrieves all services available in a specific city.","operationId":"getAllServicesInCity","parameters":[{"name":"cityId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ServiceResponse"}}}}}}},"post":{"tags":["Service Management"],"summary":"Create a new service","description":"Adds a new service to a specific city. Only admins can perform this action.","operationId":"createService","parameters":[{"name":"cityId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServiceRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ServiceResponse"}}}}}}},"/api/cities/add":{"post":{"tags":["City Management"],"summary":"Add a new city","description":"Creates a new city. Only admins can perform this action.","operationId":"addCity","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CityRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CityResponse"}}}}}}},"/api/account/register":{"post":{"tags":["Authentication"],"summary":"Register a new user","description":"Registers a new user with the provided details.","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/account/logout":{"post":{"tags":["Authentication"],"summary":"Logout user","description":"Logs out the current user by blacklisting their JWT token.","operationId":"logout","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LogoutResponse"}}}}}}},"/api/account/login":{"post":{"tags":["Authentication"],"summary":"User login","description":"Authenticates a user and returns a JWT token.","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/account/change-password":{"post":{"tags":["Authentication"],"summary":"Change password","description":"Changes the password of the current user.","operationId":"changePassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/cities":{"get":{"tags":["City Management"],"summary":"Retrieve all cities","description":"Fetches a list of all cities.","operationId":"getAllCities","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CityResponse"}}}}}}}},"/api/cities/search":{"get":{"tags":["City Management"],"summary":"Search cities","description":"Searches for cities by name and/or country.","operationId":"searchCities","parameters":[{"name":"name","in":"query","required":false,"schema":{"type":"string"}},{"name":"country","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CityResponse"}}}}}}}}},"components":{"schemas":{"ServiceCategoryRequest":{"required":["name"],"type":"object","properties":{"name":{"maxLength":100,"minLength":0,"type":"string"},"description":{"maxLength":500,"minLength":0,"type":"string"}}},"ServiceCategoryResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"CityRequest":{"required":["country","name"],"type":"object","properties":{"name":{"maxLength":100,"minLength":0,"type":"string"},"country":{"maxLength":100,"minLength":0,"type":"string"},"description":{"type":"string"}}},"CityResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"country":{"type":"string"},"description":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"EditAccountRequest":{"type":"object","properties":{"email":{"type":"string"},"fullName":{"type":"string"}}},"GetCurrentUserResponse":{"type":"object","properties":{"username":{"type":"string"},"email":{"type":"string"},"fullName":{"type":"string"},"role":{"type":"string"}}},"ChangeUserRoleRequest":{"required":["role","username"],"type":"object","properties":{"username":{"type":"string"},"role":{"pattern":"ROLE_USER|ROLE_ADMIN","type":"string"}}},"ServiceRequest":{"required":["address","categoryId","contactInfo","name"],"type":"object","properties":{"name":{"maxLength":100,"minLength":0,"type":"string"},"description":{"maxLength":500,"minLength":0,"type":"string"},"address":{"type":"string"},"contactInfo":{"type":"string"},"operatingHours":{"maxLength":1000,"minLength":0,"type":"string"},"categoryId":{"type":"integer","format":"int64"}}},"ServiceResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"categoryId":{"type":"integer","format":"int64"},"categoryName":{"type":"string"},"cityId":{"type":"integer","format":"int64"},"cityName":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"address":{"type":"string"},"contactInfo":{"type":"string"},"operatingHours":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"RegisterRequest":{"required":["email","fullName","password","username"],"type":"object","properties":{"username":{"maxLength":50,"minLength":3,"type":"string"},"email":{"type":"string"},"password":{"maxLength":2147483647,"minLength":6,"type":"string"},"fullName":{"type":"string"}}},"AuthResponse":{"type":"object","properties":{"token":{"type":"string"},"username":{"type":"string"},"email":{"type":"string"},"fullName":{"type":"string"},"role":{"type":"string"}}},"LogoutResponse":{"type":"object","properties":{"success":{"type":"boolean"}}},"LoginRequest":{"required":["password","username"],"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"ChangePasswordRequest":{"type":"object","properties":{"currentPassword":{"type":"string"},"newPassword":{"type":"string"}}}}}}